added migrations

+++ src/CorePlus.Modules/Reporting/Models/AppointmentRecord.cs
@@ -8,8 +8,8 @@ public class AppointmentRecord : BaseEntity
     public string ClientName { get; set; }
     public string AppointmentType { get; set; }
     public int Duration { get; set; }
-    public decimal Cost { get; set; }
-    public decimal Revenue { get; set; }
+    public double Cost { get; set; }
+    public double Revenue { get; set; }
     public DateTime Date { get; set; }
 
 }
